feat: drop disputed claims
This is inspired by OptimisticGovernor from UMA oSnap.
A dipusted claim does not obstruct the oracle process, but a new claim can be instantly made again.
Another idea is to make successive claims to the same claimId have a escalation factor, meaning new proposals have some multiplier on the bond amount.
This PR is equivalent to an escalation factor of 1x, ie. constant bond amount